# docs/api/openAPI.yaml
...
openapi: 3.1.0
info:
  title: Industry Core Hub Backend API
  version: 0.2.0
paths:
...
  /v1/twin-management/catalog-part-twin:
    get:
      tags:
        - Twin Management
      summary: Twin Management Get Catalog Part Twins
      operationId: >-
        twin_management_get_catalog_part_twins_v1_twin_management_catalog_part_twin_get
      parameters:
        - name: include_data_exchange_agreements
          in: query
          required: false
          schema:
            type: boolean
            default: false
            title: Include Data Exchange Agreements
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogPartTwinRead'
                title: >-
                  Response Twin Management Get Catalog Part Twins V1 Twin
                  Management Catalog Part Twin Get
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
    post:
      tags:
        - Twin Management
      summary: Twin Management Create Catalog Part Twin
      operationId: >-
        twin_management_create_catalog_part_twin_v1_twin_management_catalog_part_twin_post
      parameters:
        - name: autoCreatePartTypeInformation
          in: query
          required: false
          schema:
            type: boolean
            description: >-
              Automatically create part type information submodel if not
              present.
            default: true
            title: Autocreateparttypeinformation
          description: Automatically create part type information submodel if not present.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogPartTwinCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinRead'
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  '/v1/twin-management/catalog-part-twin/{global_id}':
    get:
      tags:
        - Twin Management
      summary: Twin Management Get Catalog Part Twin
      operationId: >-
        twin_management_get_catalog_part_twin_v1_twin_management_catalog_part_twin__global_id__get
      parameters:
        - name: global_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Global Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/CatalogPartTwinDetailsRead'
                  - type: 'null'
                title: >-
                  Response Twin Management Get Catalog Part Twin V1 Twin
                  Management Catalog Part Twin  Global Id  Get
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  '/v1/twin-management/catalog-part-twin/{manufacturer_id}/{manufacturer_part_id}':
    get:
      tags:
        - Twin Management
      summary: Twin Management Get Catalog Part Twin From Manufacturer
      operationId: >-
        twin_management_get_catalog_part_twin_from_manufacturer_v1_twin_management_catalog_part_twin__manufacturer_id___manufacturer_part_id__get
      parameters:
        - name: manufacturer_id
          in: path
          required: true
          schema:
            type: string
            title: Manufacturer Id
        - name: manufacturer_part_id
          in: path
          required: true
          schema:
            type: string
            title: Manufacturer Part Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/CatalogPartTwinDetailsRead'
                  - type: 'null'
                title: >-
                  Response Twin Management Get Catalog Part Twin From
                  Manufacturer V1 Twin Management Catalog Part Twin 
                  Manufacturer Id   Manufacturer Part Id  Get
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /v1/twin-management/catalog-part-twin/share:
    post:
      tags:
        - Twin Management
      summary: Twin Management Share Catalog Part Twin
      operationId: >-
        twin_management_share_catalog_part_twin_v1_twin_management_catalog_part_twin_share_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogPartTwinShareCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '201':
          description: Catalog part twin shared successfully
        '204':
          description: Catalog part twin already shared
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /v1/twin-management/serialized-part-twin:
    get:
      tags:
        - Twin Management
      summary: Twin Management Get All Serialized Part Twins
      operationId: >-
        twin_management_get_all_serialized_part_twins_v1_twin_management_serialized_part_twin_get
      parameters:
        - name: include_data_exchange_agreements
          in: query
          required: false
          schema:
            type: boolean
            default: false
            title: Include Data Exchange Agreements
        - name: manufacturerId
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Manufacturerid
        - name: manufacturerPartId
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Manufacturerpartid
        - name: customerPartId
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Customerpartid
        - name: partInstanceId
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Partinstanceid
        - name: van
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Van
        - name: businessPartnerNumber
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Businesspartnernumber
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SerializedPartTwinRead'
                title: >-
                  Response Twin Management Get All Serialized Part Twins V1 Twin
                  Management Serialized Part Twin Get
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
    post:
      tags:
        - Twin Management
      summary: Twin Management Create Serialized Part Twin
      operationId: >-
        twin_management_create_serialized_part_twin_v1_twin_management_serialized_part_twin_post
      parameters:
        - name: autoCreatePartTypeInformation
          in: query
          required: false
          schema:
            type: boolean
            description: >-
              Automatically create part type information submodel if not
              present.
            default: true
            title: Autocreateparttypeinformation
          description: Automatically create part type information submodel if not present.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SerializedPartTwinCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinRead'
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  '/v1/twin-management/serialized-part-twin/{global_id}':
    get:
      tags:
        - Twin Management
      summary: Twin Management Get Serialized Part Twin
      operationId: >-
        twin_management_get_serialized_part_twin_v1_twin_management_serialized_part_twin__global_id__get
      parameters:
        - name: global_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Global Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SerializedPartTwinDetailsRead'
                  - type: 'null'
                title: >-
                  Response Twin Management Get Serialized Part Twin V1 Twin
                  Management Serialized Part Twin  Global Id  Get
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /v1/twin-management/twin-aspect:
    post:
      tags:
        - Twin Management
      summary: Twin Management Create Twin Aspect
      operationId: twin_management_create_twin_aspect_v1_twin_management_twin_aspect_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwinAspectCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinAspectRead'
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /v1/twin-management/serialized-part-twin/share:
    post:
      tags:
        - Twin Management
      summary: Twin Management Share Serialized Part Twin
      operationId: >-
        twin_management_share_serialized_part_twin_v1_twin_management_serialized_part_twin_share_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SerializedPartTwinShareCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '201':
          description: Catalog part twin shared successfully
        '204':
          description: Catalog part twin already shared
        '400':
          description: Invalid input provided. Please check your request and try again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Access denied. You do not have permission to perform this action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: Catalog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Catalog part already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: >-
            Bad Gateway - The server received an invalid response from the
            upstream server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: Service unavailable. Please try again later.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
...
components:
  schemas:
...
    CatalogPartTwinCreate:
      properties:
        globalId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Globalid
          description: >-
            Optionally the Catena-X ID / global ID of the digital twin to
            create. If not specified, a new UUID will be created automatically.
        dtrAasId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Dtraasid
          description: >-
            Optionally the shell descriptor ID ('AAS ID') of the digital twin in
            the Digital Twin Registry. If not specified, a new UUID will be
            created automatically.
        idShort:
          anyOf:
            - type: string
            - type: 'null'
          title: Idshort
          description: >-
            Optionally the idShort of the digital twin in the Digital Twin
            Registry. If not specified, a default value 'Twin-{globalId}' will
            be used.
        manufacturerId:
          type: string
          title: Manufacturerid
          description: The BPNL (manufactuer ID) of the part to register.
        manufacturerPartId:
          type: string
          title: Manufacturerpartid
          description: The manufacturer part ID of the part.
      type: object
      required:
        - manufacturerId
        - manufacturerPartId
      title: CatalogPartTwinCreate
...
    CatalogPartTwinShareCreate:
      properties:
        businessPartnerNumber:
          type: string
          title: Businesspartnernumber
          description: >-
            The business partner number of the business partner with which the
            catalog part is shared.
        manufacturerId:
          type: string
          title: Manufacturerid
          description: The BPNL (manufactuer ID) of the part to register.
        manufacturerPartId:
          type: string
          title: Manufacturerpartid
          description: The manufacturer part ID of the part.
      type: object
      required:
        - businessPartnerNumber
        - manufacturerId
        - manufacturerPartId
      title: CatalogPartTwinShareCreate
...
    ErrorDetail:
      properties:
        status:
          type: integer
          title: Status
        message:
          type: string
          title: Message
      type: object
      required:
        - status
        - message
      title: ErrorDetail
...